Extract language features into typescript server plugin (#89)
This has been a long time coming. Extracted most of the previous features into a typescript server plugin, which means that the tooling for Imba now runs in the same worker / environment as the tooling for js and typescript. It provides a ton of additional features.

This will be the very first release using this approach, so expect bugs and stability issues. The goal is to get back to more frequent updates now that the groundwork has been laid. Will also write a post showing some of the new features and also listing the known limitations and issues that we're aiming to resolve shortly.